Indicate that `main.tex` is the LaTeX "master" file.
When a LaTeX project contains multiple source files
(e.g. for separate chapters) that are "included" in
a master file, Vim's Latex-Suite needs to know
which file is the master file. The master file
is specified via an *empty* file with the following
naming convention:

    *.latexmain

where * is the name of the LaTeX master file.
In our case, `main.tex` is our master file, so
we have created an empty file `main.tex.latexmain`
in the same directory as the rest of the project.

More information on this topic is available here:

http://vim-latex.sourceforge.net/documentation/latex-suite/latex-project.html