Do not ask for user input upon each LaTeX compilation.
Vim's LaTeX-Suite does not like non-continuous compilation,
such as asking for user input with `\typein`, which
was being used in the MIT template to ask which files
to include in the compilation.

Instead, we have simply hard coded the files to be compiled.
The user can still fairly easily/dynamically change the files
by specifying them in `main.tex` prior to compilation.