`appb` is now its own directory.
In the future, each chapter/appendix will be contained
in its own directory, containing the source LaTeX file,
a `figs` file for the figures to be included in the
chapter, etc. Hopefully, this makes each chapter into
a more organized, self-standing unit.

Additionally, the LaTeX auxiliary compilation files
are also stored in their respective directories,
which will reduce clutter in the top directory
of the thesis repo.